Placeholder - Add offset property to PlaceholderState.
Chip and Card background placeholder painters need to have the offset of the component they are drawing passed into them in order to allow the coordination of the placeholder effects.
Add this to the state and populate it from the Modifier.placeholderShimmer.

Bug: 253667329
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Relnote: "We have corrected a UI bug in the placeholder wipeOff effect where Chip and Card backgrounds were wiping off slightly early due to not taking the components position on screen into account."
Change-Id: I2c7cb2ed5d27ec425cca21dca7bf74cab3d3d854