Convert DeleteOrUpdateMethodAdapter to XPoet
Convert the simple path for @Update, @Delete and @Upsert shortcut methods to generate Kotlin, i.e. the path that uses the instant binders. @Upsert was mostly done since it shares its adapter with @Insert but the binder needed converting and since it uses update too it had to be done in at the same time.

Bug: 127483380
Test: KotlinCodeGenTest
Change-Id: I4b606c3fa71a02c6f112a836590d31c17fb4f2e1