[CustomTabs] Add new parameters to CustomTabsCallback#onActivityResized
Changed the CustomTabsCallback#onActivityResized behavior to be called for all activity resizes. Removed the 'size’ parameter and added the height and width parameters to account for both changes in the activity's height and width.

Renamed EXTRA_ACTIVITY_RESIZE_BEHAVIOR to EXTRA_ACTIVITY_RESIZE_HEIGHT_BEHAVIOR to better reflect that it is height specific.

Test: ./gradlew browser:browser:connectedAndroidTest and ./gradlew browser:browser:assemble
Relnote: "Changed the CustomTabsCallback#onActivityResized behavior and added new parameters to it.
Renamed EXTRA_ACTIVITY_RESIZE_BEHAVIOR to EXTRA_ACTIVITY_RESIZE_HEIGHT_BEHAVIOR to better reflect that it is height specific."

Change-Id: Ic864e35d84188422f9a40d3535d559dcac23760c