Avoid crashing on ClassCastException in SavedStateHandle
When using SavedStateHandle, if you attempt to retrieve a value with the
incorrect class type, it is casted to the correct class type and ends up
throwing a ClassCastException.

Instead of throwing, it would be nicer if we caught the exception,
removed the value from the handle since that was not the expected value
anyway, and return a null. This way we prevent the crash and help get
the handle closer to the expected state for future interactions.

RelNote: "`SavedStateHandle` no longer crashes with a
`ClassCastException` when calling `get()` with the incorrect class
type."
Test: ./gradlew bOS
Bug: 213646066

Change-Id: I6ae7c23cd7e2ac1ec35bfc6b38aef973be1040a0