Split ArraySet into separate Jvm and Native implementations for toArray compatibility
It is required to provided a separate Java implementation of `.toArray` for
ArraySet for compatibility, due to:
  1. Ignoring nullity of the generic type T to maintain the behavior of setting the n+1'th item to null
  2. Allowing .toArray(array: Array<T>) to be called without making T a reified type
  2. Delegate the array copy implementation to System.arraycopy which is implemented natively on JVM

Instead of providing a less-performance version for native / non-jvm platforms,
it was removed in this CL because it was always only ever needed for
compatibility. Users should instead prefer to follow the Kotlin stdlib standard
of using `.toTypedArray` or one of its primitive relatives, e.g., `.toIntArray`.

Relnote: "`ArraySet.toArray` is now only available on JVM as it was always only
  included for compatibility with Java. Kotlin users should generaly use the
  stdlib's `.toTypedArray` helpers instead, which provides the exact same
  functionality with added type safety."
Test: :collection:collection:test
Change-Id: I2c59be76341e284936d6da16a3feb582d1f66c51