Put nested fragment without childFragmentManager behind flag
When nesting fragments, the child fragment should
be added using the parent's childFragmentManager.
However, since we previously allowed this
functionality, we want to put its removal behind
a flag in FragmentStrictMode.

RelNote: "The functionality of nested fragments
without the use of the parent's
`childFragmentManager` is now behind a flag in
`FragmentStrictMode`, which can be enabled to
check for violations."
Bug: 249299268
Test: ./gradlew checkApi

Change-Id: I725210f957b9be5875f7c0eadaaa992deda6614a