Creating version consistent with lower API levels
This will ensure it works for 21 and higher, and for < 21, it will have
a flashing transition artifact that will not impact utilization. This is
due to certian styles not existing prior to API 21, and due to the
deprecated nature of fragment objects after API 24 and certain
'getFragment()' calls from fragment managers after API 26. Also tagging
in some improvements to the JSON parsing, will continue those in more
depth in a future change.

Bug: 261232711
Test: JUnit Tests will be chained in
Change-Id: I115f4bd93f62730670ed4814cd8797a921a1a1b0