Define offsets in lazy animate item placement tests in floats
It makes the test results more reliable. There are multiple ways to define an offset based on the fraction. First one is
startPosition + ((endPosition - startPosition) * fraction).roundToInt()
And the second one is:
endPosition - ((endPosition - startPosition) * fraction).roundToInt()
Lets say startPosition is 0 and startPosition is 9. Then for the first expression we have
0 + ((9 - 0) * 0.5).roundToInt() = 4.5.roundToInt() = 5
And for the second one we have
9 - ((9 - 0) * 0.5).roundToInt() = 9 - 4.5.roundToInt() = 9 - 5 = 4
But if instead we will be using floats for calculations and only round the result to int right before doing the assertion both expressions will produce 4.5f which we round to 5.

Test: run those tests
Change-Id: I9e590a066b2e41b6e00a293e7a7fcde09a312b69