Fix behavior when using Intrinsics with ConstraintLayout & MotionLayout
Intrinsics trigger secondary measure calls without recomposing.

ConstraintLayout was not updating its size properly since the assumption
was that it was only necessary if the constraints changed or the content
is recomposed, none of which happen with Intrinsic measurement.

MotionLayout would attempt to call measure on outdated Measurables,
causing a crash since the mapping of Widgets to Measurables is only
done if the content recomposed. The fix was also applied to
ConstraintLayout since it's a safer pattern.

Fixes: 220527863
Test: added respectective test cases
Change-Id: I487ae6fcc36342cee7fe38428971c27500be0b23