wearos-whs: Remove default throw Exception impl in ExerciseClient.updateExerciseTypeConfigAsync
According to feedback on beta API change request, we should remove the
default impl of throwing Exception in
ExerciseClient.updateExerciseTypeConfigAsync, since this approach
currently makes a compile time error into a runtime error which makes
the developer ergonomics non-ideal.

All developers are expected to use
ServiceBackedExerciseClient.updateExerciseTypeConfigAsync, where we have
well provided an implementation. So the throwing of Exception should not
be invoked at all, it deserves removing this Exception in the function
body and signature, as well as the kdoc of "@see ServiceBackedExerciseClient"
implementation detail that we don't need to direct developers to (but we
previously introduced together with the "throw Exception" change)

Since this cl requires API change, and this cl is already part of a
filed beta API change request, we are using --overridePolicyWithApproval=260745756 flag to pass the updateApi command

Relnote: Remove default implementation of throwing Exception
Test: ./gradlew :health:health-services-client:assemble --info
Test: ./gradlew :health:health-services-client:test
Bug: 260745756

Change-Id: Id947f2552f3c6e4e9abd981f09f6922db3f4fa7d