Compatibility test for Person
We have added internal schema types for Person and ContactPoint in the
Framework from T. And we do want our clients to use this Person and
ContactPoint java classes to read the contact data based on those two
internal schema types.

Add a test to make sure any changes to the Person and ContactPoint
classes won't break this use case.

Test: ./gradlew appsearch:appsearch-builtin-types:connectedCheck

Change-Id: Iab3f102a3982fa2aba146b6b7ddc36702791cc6c