Fixes flaky testSeekBarPreferenceChangeListenerWithContinuousUpdates
Swipes were either being injected too fast / the device was too slow to see all the intermediate move events, causing flaky tests. The goal of this test is to ensure that the SeekBarPreference is updating during a swipe, so as long as we see more than one update we know this to be true, even if there is flakiness with the injected swipe in the test environment.
Test: SeekBarPreferenceTest
Fixes: b/213245076

Change-Id: I27cdbc53f1ca45523988d1e3c74ebb52a3ef0ab5