[Extensions] Fixed surface abandoned crash when pause/resume too quickly on Samsung Z Fold2
This cl is similar as the reverted CL aosp/2207335 as I think it
is still necessary to phase in to make the stability on par
with Camera2 Extensions. We enable the workaround only
on Samsung devices.

In camera2 Extensions, all camera event will be passed through
AIDL which has the extra delay so it avoids the crash without
any efforts.

Root cause:
On Samsung Z Fold2, when Extensions are enabled, if pause / resume
camera too quickly, it might cause surface abandoned crash.

Samsung will start the initialization at onEnableSession,
and release the resource at onDisableSession. The issue is
caused by onDisableSession being invoked too early (within
50ms) and could destroy the ImageWriter surface while it
is still used in the initialization.

The workaround is to ensure onDisableSession being called in
at least 100ms after onEnableSession. It will sleep the
camera thread for a while to achieve this. Though the CL
could cause some delay when shuting down the camera,
it is not the normal user operation to close the camera
and open the camera quickly within 100ms. This normally
happens in auto testing. So the impacts for the user
is almost none.

Relnote: "Fixed the Extensions crash that happens on some Samsung
devices when pausing/resuming the app too quickly"
Bug: 264502486
Bug: 245226085
Test: extensions/OpenCloseCaptureSessionStressTest, OnEnableDisableSessionDurationCheckTest

Change-Id: Iaca262a499d7e8e72c5072ea7063943dd291b4f2