Reformat files under camera/camera-camera2-pipe with ktfmt
This change helps keep file formatting consistent via the ktfmt tool.
In order to match the existing AndroidX AndroidStudio project settings
and default formatting, the ktfmt tool was invoked using:

ktfmt --dropbox-style camera/camera-camera2-pipe

Which preserves the 4 space indenting and has the minimal amount of
formatting impact to existing users. It is still required that CLs
continue to pass the ktlint tool.

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest\
 :camera:integration-tests:camera-testapp-camera2-pipe:testDebugUnitTest

Change-Id: I8af401afa66963e22f434eab3fb82fb924b8a4b3