Remove unsupported vector properties for the switch on/off animations for api levels < 24. The `android:fillColor` property accepts gradient colors api24+ only. Similarly, `android:fillType` is supported in API24+ only. See https://developer.android.com/reference/android/graphics/drawable/VectorDrawable.html.
It was failing to load the vector drawable in API23 causing the widget to show "Problem loading the widget" error. Other animations didn't seem to use gradients, so seem to load fine.
Alternatively, we could have used srcCompat, but with that, the existing switch vector as is, didn't load properly showing just plain single color. Moreover, we don't depend on support library currently, so adding a dependency seems unnecessary for this.

Bug: 266436897
Test: CompoundButtonAppWidget renders fine across different backport versions
Change-Id: Ibb40371b525dbe521daf54a331a1e8181221b190