Fix custom layout constraints in PickerGroup
PickerGroup relies on custom layout to center the items inside itself. When `Modifier.fillMaxSize()` was applied on the PickerGroup, it was setting the minHeight and minWidth same as that of maxHeight and maxWidth. This caused the child composables to take up the whole space leading to a single picker taking up all the space and hiding the other pickers.

This change resets the incoming constraints from the parent composable for min width and height, allowing the child composables to take only the space that is required. Apart from that, the custom layout does not need height more than the max height of child composables which will make the layout compact in terms of its size while including all the children properly.

Relnote: """
Update PickerGroup API to allow propagating the min constraints to the composable optionally. When set as true, the minimum constraints passed from the parent composable will be allowed on the PickerGroup. If set as false, PickerGroup will reset the minimum constraints.
"""

Bug: 264147685
Test: Manual testing

Change-Id: I3e046ae05db206b34b106c5601891de0e3913e0a