Changing Slider API version
Reverting the Slider API with no custom thumb and track to be stable. Removing Slider overload methods that are covered by the new experimental Slider API that contains custom thumb and track. Changing the RangeSlider API with custom thumbs and track to be experimental. Promoting RangeSlider API with no custom thumbs and track to be stable.

Bug: 269327141
Test: N/A
Relnote: reverting some Slider API versions and removing redundant Slider overloads that are covered by the new experimental Slider with custom thumb and track. Promoting RangeSlider to be stable.
Change-Id: Ie8fbdf564fa3e2a83ff891a79bc0baa5c12e586d