Fix dynamicNavHost with new FragmentNavigator
When using Dynamic Navigation, if you navigate after inflating with a
FragmentContainerView, the DynamicNavigator crashes because it attempts
to register an onBackChangedListener before the navigator is attached.

Instead of attaching an onBackChangedListener to the FragmentManager in
the init block, do it in onAttach().

RelNote: "Fixed a crash when attempting to navigate while using Dynamic
Navigation."
Test: tested in sample App
Bug: 268360479

Change-Id: I3ee29ea33f2cc7f7221b79d4d17399c725a6a4df