Handle cases where read fails after init
This CL adds 2 new tests wrt how we handle read errors after
failures.

The first one, `readFailsAfter_failedUpdate` ensures that
updateData failures do not affect existing readers or cached
value.

The second one, `readFailsAfter_successfulUpdate` ensures that
if a read error happens when we have cached value, it is reflected
in the cached data and dispatched to existing readers.

To fix this, i needed 2 changes:

* handleRead wasn't handling exceptions, now it does
* data flow was skipping non-data values when the initial cached
value is Data. This is wrong as it will cause the flow to never
finish (e.g. it cannot transition to Final from Data or ReadException
from data). I've changed it to handle these cases properly.

To be able to test this, I opened up an internal API that lets us
access the coordinator from tests.

Note that, the case fixed by this test could've never happened in
SPDS, can only happen in MPDS as SPDS always uses cached value.
Nevertheless, in the new combined implementation, testing it with
SPDS is easier. Furthermore, it was already broken in MPDS before the
merged implementation CL (aosp/2439282 for its repr).

Fixes: 269533609
Bug: 267822693
Bug: 269772127
Test: SingleProcessDataStoreTest
Change-Id: I431eb63c5c8a6c570f3a5c27f4ce50024527a74f