Simplify AndroidView's nestedScroll Modifier usage
This CL moves the nestedScroll modifier from the AndroidView composable
into the AndroidViewHolder class. We can't do this without nestedScroll
migrating to Modifier.Node first, because we stop materializing it as a
result of this change.

This should also fix a known issue with the reusable AndroidView
composable where nested scrolling may not work correctly after a View
instance is reused.

Test: ./gradlew compose:ui:ui:cAT
Change-Id: I2519fc3206d4294a9899c301e2c1b64de899865f