Make Controller classes stateless.
Convert SplitController, RuleController, and ActivityEmbeddingController
to be stateless.  All the state should be managed by EmbeddingBackend.
A stateless controller will also make the corresponding test rule
correct. We had a bug where if you used a singleton it would capture a
mocked backend and if you ran two tests the second would fail since you
use a new backend.  Now the backend is injected on each run.
Move test to be unit test.

Bug: 269360339
Test: pass existing tests.
Change-Id: I3d59660d48ee2d24a9bf824034a2da05d0f96942