Migrate to public AGP APIs
Except for BaseExtension, which is required to access the properties
needed to find AIDL and the framework AIDL file. Also, I copied the
DependencyFileProcessor out of AGP and fixed formatting / suppressed
ErrorProne warnings.

Long-term, we might want to simplify some of the cloned classes. For now,
though, I wanted to minimize the amount of review necessary.

Bug: 270155261
Test: existing Stable AIDL test suite
Change-Id: Idbdcdcda400a995b9331d557587441c6dd782235